move PassManagerContext to a file
