+++ qiskit/transpiler/passmanager.py
@@ -15,7 +15,6 @@
 
 from functools import partial
 from collections import OrderedDict
-from time import time
 
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.converters import circuit_to_dag, dag_to_circuit
@@ -24,6 +23,7 @@ from .propertyset import PropertySet
 from .basepasses import BasePass
 from .fencedobjs import FencedPropertySet, FencedDAGCircuit
 from .exceptions import TranspilerError
+from .passcontext import PassManagerContext
 
 
 class PassManager():
@@ -205,37 +205,6 @@ class PassManager():
         for pass_ in self.working_list:
             ret.append(pass_.dump_passes())
         return ret
-
-
-class PassManagerContext:
-    """ A wrap around the execution of a pass."""
-
-    def __init__(self, pm_instance, pass_instance):
-        self.pm_instance = pm_instance
-        self.pass_instance = pass_instance
-        self.start_time = None
-
-    def __enter__(self):
-        if self.pm_instance.log_passes:
-            self.start_time = time()
-
-    def __exit__(self, *exc_info):
-        if self.pm_instance.log_passes:
-            end_time = time()
-            raw_log_dict = {
-                'name': self.pass_instance.name(),
-                'start_time': self.start_time,
-                'end_time': end_time,
-                'running_time': end_time - self.start_time
-            }
-            log_dict = "%s: %.5f (ms)" % (self.pass_instance.name(),
-                                          (end_time - self.start_time) * 1000)
-            if self.pm_instance.property_set['pass_raw_log'] is None:
-                self.pm_instance.property_set['pass_raw_log'] = []
-            if self.pm_instance.property_set['pass_log'] is None:
-                self.pm_instance.property_set['pass_log'] = []
-            self.pm_instance.property_set['pass_raw_log'].append(raw_log_dict)
-            self.pm_instance.property_set['pass_log'].append(log_dict)
 
 
 class FlowController():
